backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb==0.5.3",
    "fastapi==0.111.0",
    "httpx==0.27.0",
    "langchain==0.2.6",
    "langchain-chroma==0.1.1",
    "langchain-community==0.2.6",
    "langchain-groq>=0.1.10",
    "langchain-huggingface>=0.0.3",
    "langchain-openai==0.1.13",
    "pydantic>=2.12.5",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "uvicorn==0.30.1",
]
